Membre maj

+++ ncdz-boot/src/main/java/fr/formation/dao/IDAOMembre.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import fr.formation.model.Membre;
 
-public interface IDAOMembre extends JpaRepository<Membre, Integer> {
+public interface IDAOMembre extends JpaRepository<Membre, Long> {
 
 }
